% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_decimal_points_to_dot.R
\name{all_decimal_points_to_dot}
\alias{all_decimal_points_to_dot}
\title{Convert given decimal character to dot}
\usage{
all_decimal_points_to_dot(df, convert_decimal_character = ",", trim_ws = FALSE)
}
\arguments{
\item{df}{A data.frame or tibble.}

\item{convert_decimal_character}{Specify the decimal point character that shall be converted to a dot.}

\item{trim_ws}{Trim white space from both ends before trying conversion?}
}
\value{
A similar object as \code{df}, but with potentially replaced decimal characters.
}
\description{
Convert specified decimal character in numeric-like character to dot.
}
\details{
Has safeguards to only convert numeric-like characters - any non-numeric character will prevent conversion.
}
\examples{
df <- data.frame(x1 = c("1,1", "2,2", " 3,3", "4,4 ", "1,1a"))
all_decimal_points_to_dot(df)

# also trim white space
all_decimal_points_to_dot(df, trim_ws = TRUE)
}
