% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_to_character.R
\name{all_to_character}
\alias{all_to_character}
\title{Convert all columns to character}
\usage{
all_to_character(df)
}
\arguments{
\item{df}{A data.frame or tibble which columns shall be converted.}
}
\value{
A similar object as \code{df}, but all columns are of class \code{character()}.
}
\description{
Convert all columns of the provided object to \code{character}.
}
\examples{
df <- data.frame(a = c(1, 2, 3), b = c("a", "b", "c"))
df

all_to_character(df)
}
