% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rowwise.R
\name{apply_rowwise}
\alias{apply_rowwise}
\title{Apply a function itemwise to two vectors}
\usage{
apply_rowwise(
  vec1,
  vec2,
  apply_fun,
  wrap_fun_vec1 = NULL,
  wrap_fun_vec2 = NULL
)
}
\arguments{
\item{vec1}{Atomic vector 1.}

\item{vec2}{Atomic vector 2.}

\item{apply_fun}{The function that should be applied to both elements from \code{vec1} and \code{vec2} at the same index. Can be a function or a function name.}

\item{wrap_fun_vec1}{Optionally, a function that shall be applied to every \code{vec1} element before applying \code{apply_fun}. Can be a function or a function name.}

\item{wrap_fun_vec2}{Optioanlly, a function that shall be applied to every \code{vec2} element before applying \code{apply_fun}. Can be a function or a function name.}
}
\value{
A result of the same length as the input vectors.
}
\description{
Apply a function rowwise to two same length vectors, e.g. comparison functions.
Optionally apply wrap functions to each item before the apply function.
Functions can also be function names (e.g. "sum" or "+").
}
\examples{
# Can be useful for comparing vectors of potentially different data types.
vec1 <- c(1:5)
vec2 <- c(0:2, "a", "b")

apply_rowwise(vec1, vec2, apply_fun = ">", wrap_fun_vec2 = try_numeric)

}
