% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when_paste.R
\name{case_when_paste}
\alias{case_when_paste}
\title{case_when variant that pastes results of matched formulas.}
\usage{
case_when_paste(..., .default = "", .sep = "")
}
\arguments{
\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A sequence of two-sided
formulas. The left hand side (LHS) determines which values match this case.
The right hand side (RHS) provides the replacement value.

For \code{case_when()}:
\itemize{
\item The LHS inputs must be logical vectors. For backwards compatibility,
scalars are \link[vctrs:theory-faq-recycling]{recycled}, but we no longer
recommend supplying scalars.
\item The RHS inputs will be \link[vctrs:theory-faq-coercion]{cast} to their common
type, and will be \link[vctrs:theory-faq-recycling]{recycled} to the common
size of the LHS inputs.
}

For \code{replace_when()}:
\itemize{
\item The LHS inputs must be logical vectors the same size as \code{x}.
\item The RHS inputs will be \link[vctrs:theory-faq-coercion]{cast} to the type of
\code{x} and \link[vctrs:theory-faq-recycling]{recycled} to the size of \code{x}.
}

\code{NULL} inputs are ignored.}

\item{.default}{The value used when all of the LHS inputs return either
\code{FALSE} or \code{NA}.
\itemize{
\item If \code{NULL}, the default, a missing value will be used.
\item If provided, \code{.default} will follow the same type and size rules as the
RHS inputs.
}

\code{NA} values in the LHS conditions are treated like \code{FALSE}, meaning that
the result at those locations will be assigned the \code{.default} value. To
handle missing values in the conditions differently, you must explicitly
catch them with another condition before they fall through to the
\code{.default}. This typically involves some variation of \code{is.na(x) ~ value}
tailored to your usage of \code{case_when()}.}

\item{.sep}{The string separator for the different results when being combined.}
}
\value{
A vector with the same size as the inputs.
}
\description{
Drop-in replacement for \link[dplyr:case-and-replace-when]{case_when()}, when you want combined output from the provided formulas.
Evaluates each formula separately, then combines the result per entry.
Drops \code{NA} results before pasting - \code{NA} should only be supplied in \code{.default} if necessary.
}
\examples{
x <- data.frame(a = c(1, 2, 3), b = c(1, 2, 4))
x

dplyr::mutate(x,
  class = case_when_paste(
    a == b ~ "equal",
    a == 1 ~ "special",
    .default = "not equal",
    .sep = " - "
  )
)
}
