% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_one_hot.R
\name{check_one_hot}
\alias{check_one_hot}
\title{Check one-hot encoding}
\usage{
check_one_hot(df, cols, check_value = 1, allow_no_selection = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble}.}

\item{cols}{Which columns of \code{df} shall be checked?}

\item{check_value}{Which value represents a \verb{positive selection}?}

\item{allow_no_selection}{Shall rows without \code{check_value} be valid or not?}
}
\value{
A logical vector indicating for every row of \code{df} if the selected columns follow a one-hot encoding scheme.
}
\description{
Check if specified columns follow a one-hot coding scheme, i.e. only one "positive" value per row across the columns.
}
\details{
You can specify if rows without selection, i.e. rows without "positive" value, should be regarded as valid.
Does not check the format of the other "negative" values.
}
\examples{
df <- data.frame(a = c(1, 0, 0), b = c(0, 0, 1), c = c(0, NA, 1))
df

# every row must have a positive selection
check_one_hot(df, cols = colnames(df))

# also allow empty rows
check_one_hot(df, cols = colnames(df), allow_no_selection = TRUE)

# also works with other "positive" values
df <- data.frame(a = c("one", "two"), b = c("two", "oneone"))
df

check_one_hot(df, cols = colnames(df), check_value = "one")

}
