% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trailing.R
\name{check_trailing}
\alias{check_trailing}
\title{Check for trailing elements}
\usage{
check_trailing(vec, trail_value)
}
\arguments{
\item{vec}{The to-be-checked vector.}

\item{trail_value}{The trailing value to check for.}
}
\value{
A single logical if the provided trailing is fulfilled.
}
\description{
Check if the provided vector has a consistently trailing element.
If a certain trailing value has been reached, all subsequent vector elements should be this value.
}
\details{
If the trailing value does not occur in the vector, returns \code{TRUE}.
}
\examples{

# 3 is trailing
vec <- c(1, 2, 3, 3, 3)
check_trailing(vec, trail_value = 3)

# c is not trailing (interrupted by b)
vec <- c("b", "c", "b", "c", "c")
check_trailing(vec, trail_value = "c")

# trailing value does not occur - trailing rule is not violated
check_trailing(vec, trail_value = "d")

# also works with NA
check_trailing(c(1, NA, NA), trail_value = NA)

}
