% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trailing_rowwise.R
\name{check_trailing_rowwise}
\alias{check_trailing_rowwise}
\title{Check rowwise for trailing elements}
\usage{
check_trailing_rowwise(df, cols, trail_value)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble}.}

\item{cols}{Which columns of \code{df} shall be checked rowwise?}

\item{trail_value}{Which trailing value to check for?}
}
\value{
A logical vector of the same size as \code{df} has rows.
}
\description{
Check rowwise for the provided data.frame columns if a value is trailing.
}
