% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_hidra.R
\name{connect_hidra}
\alias{connect_hidra}
\title{Connect to the HIDRA database}
\usage{
connect_hidra(system = "mac")
}
\arguments{
\item{system}{'mac' or 'windows'}
}
\value{
Creates a database connection
}
\description{
Connect to the HIDRA database
}
\examples{
 \dontrun{connect_hidra("mac")}
}
