% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_triage_project.R
\name{create_triage_project}
\alias{create_triage_project}
\title{Create Data Triage Project}
\usage{
create_triage_project(path)
}
\arguments{
\item{path}{where to create the project}
}
\description{
Create Data Triage Project
}
