% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_culprit.R
\name{find_culprit}
\alias{find_culprit}
\title{Find not-fitting element (culprit) in vector}
\usage{
find_culprit(
  vec,
  should_be = c("majority", "numeric", "double", "integer", "character", "logical"),
  return_index = FALSE,
  na.rm = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{vec}{Data vector to find the not-fitting element in.}

\item{should_be}{Expected data type of \code{vec}. \code{Numeric} is indifferent if double or integer is provided. If set to \code{majority}, expect the data type of most of the elements.}

\item{return_index}{Return index of culprit elements rather than their value.}

\item{na.rm}{Do not treat \code{NA} values as culprits, i.e. discard them from the culprit return.}

\item{verbose}{If guessed data type should be printed if \code{should_be} is \code{majority}.}

\item{...}{
  Arguments passed on to \code{\link[readr:parse_guess]{readr::guess_parser}}
  \describe{
    \item{\code{na}}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}
    \item{\code{locale}}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[readr:locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
  }}
}
\value{
A named vector, with either the element values (default) or their index (\code{return_index}) and the guessed data types as names.
}
\description{
Find the element(s) in a vector, whose data type potentially deviates from the majority data type or a specified one.
}
\details{
This function can be used to find the culprit element, which prevents clean data type conversion.
Utilizes \code{\link[readr:parse_guess]{readr::guess_parser()}} to guess type of every vector element.
User can specify which data type the vector \code{should_be}, or simply rely on the majority guess.
By default does not differentiate between double or integer values. Set \code{should_be} to \code{integer} to enforce this differentiation.
Ties in the majority guess are resolved lexicographically by data type name, i.e. \code{character} > \code{double} > \code{integer} > \code{logical}.
If \code{should_be} is \code{numeric}, the data type is interpreted as \code{double} for lexicographic sorting.
}
\examples{
long_data <- c(1, 2.2, "4,", 5)

# standard type conversion would convert "4," to NA
\dontrun{
as.numeric(long_data)
}

# better for maximizing information content: find culprit element and treat it appropriately
find_culprit(long_data, should_be = "numeric")
as.numeric(gsub(",", "", long_data))
}
