% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_row_values.R
\name{move_row_values}
\alias{move_row_values}
\title{Move specific values to other column}
\usage{
move_row_values(
  df,
  select_rows = colnames(df)[1],
  select_rows_values = NULL,
  from_col,
  to_col,
  clear_value = NA
)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble}.}

\item{select_rows}{Which rows of \code{df} shall be selected for moving? Can be indices, logical values or a column name of \code{df}.}

\item{select_rows_values}{If a column name was supplied in \code{select_rows}, specify the column values of the rows you want to select.}

\item{from_col}{From which columns of \code{df} shall the values of selected entries be moved?}

\item{to_col}{To which columns of \code{df} shall the values of selected entries be moved?}

\item{clear_value}{With what value shall the moved values in \code{from_col} be replaced?}
}
\value{
An object with same shape and class as \code{df}.
}
\description{
Move specific values in a \code{data.frame} or \code{tibble} to other column(s).
Specify which rows shall be affected by either providing row indices or a logical vector to \code{select_rows}.
A column name can also be supplied to \code{select_rows}, then the matching values of rows have to be supplied in \code{select_rows_values}.
It is possible to specify more than one \code{from_col} and \code{to_col} - but always the same number of columns.
Moved values in \code{from_col} are replaced with \code{clear_value}.
}
\examples{
df <- data.frame(
  "id" = letters[1:5], "a" = c(5:1), "b" = c(5:4, NA, 2:1),
  "c" = c(1:5), "d" = c(1:2, 2, 4:5)
)
df

# move for row 3, column a to b - replace with NA
df_mov <- move_row_values(df, select_rows = 3, from_col = "a", to_col = "b", clear_value = NA)
df_mov

# move multiple columns and rows
df_mov <- move_row_values(df,
  select_rows = "id", select_rows_values = c("c", "d"),
  from_col = c("a", "c"), to_col = c("b", "d"), clear_value = 3
)
df_mov

# switching values is also possible e.g. by moving from a to b and b to a.
df_mov <- move_row_values(df,
  select_rows = c(FALSE, FALSE, TRUE),
  from_col = c("a", "b"), to_col = c("b", "a"), clear_value = NA
)
df_mov
}
