% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reguess_coltypes.R
\name{reguess_coltypes}
\alias{reguess_coltypes}
\title{Re-convert character columns in existing data frame}
\usage{
reguess_coltypes(df, convert_comma_to_dot = FALSE, silent = TRUE, ...)
}
\arguments{
\item{df}{A data.frame or tibble which character columns class shall be re-guessed.}

\item{convert_comma_to_dot}{Shall numeric-like characters with a comma as decimal point like '1,2' be converted to '1.2' before column type guessing?}

\item{silent}{Suppress column specification message from \code{\link[readr:type_convert]{readr::type_convert()}} column type guessing.}

\item{...}{
  Arguments passed on to \code{\link[readr:type_convert]{readr::type_convert}}
  \describe{
    \item{\code{col_types}}{One of \code{NULL}, a \code{\link[readr:cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, column types will be imputed using all rows.}
    \item{\code{na}}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}
    \item{\code{trim_ws}}{Should leading and trailing whitespace (ASCII spaces and tabs) be trimmed from
each field before parsing it?}
    \item{\code{locale}}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[readr:locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
    \item{\code{guess_integer}}{If \code{TRUE}, guess integer types for whole numbers, if
\code{FALSE} guess numeric type for all numbers.}
  }}
}
\value{
A similar object as \code{df}, but column types potentially have been changed.
}
\description{
Convenience wrapper around \code{\link[=all_decimal_points_to_dot]{all_decimal_points_to_dot()}} and \code{\link[readr:type_convert]{readr::type_convert()}}.
Optionally converts numeric-like characters with a comma as decimal point (as used in e.g. Germany) to dots,
then performs a re-guessing of column types of all character columns.
}
\examples{
# create data frame of characters
df <- data.frame(a = c(1, 2, 3), b = c(TRUE, FALSE, NA), c = factor(c(1, 2, 1))) |>
  all_to_character()
df

reguess_coltypes(df)

# convert column b to numeric if convert_comma_to_dot is set.
df <- data.frame(a = c("a,b", "1,b", "c,0"), b = c("1,1", "1,2", "1,3"))
reguess_coltypes(df, convert_comma_to_dot = TRUE)
}
