% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_multi.R
\name{relocate_multi}
\alias{relocate_multi}
\title{Relocate multiple columns}
\usage{
relocate_multi(df, move = NULL, before = NULL, after = NULL)
}
\arguments{
\item{df}{A data.frame or tibble with columns that shall be relocated.}

\item{move}{The columns that shall be moved. Can be column names, indices, <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> or <\code{\link[rlang:args_data_masking]{data-masking}}> expressions.}

\item{before}{The columns, where the respective \code{move} column shall be placed \strong{before}. Can be column names, indices, <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> or <\code{\link[rlang:args_data_masking]{data-masking}}> expressions.}

\item{after}{The columns, where the respective \code{move} column shall be placed \strong{after}. Can be column names, indices, <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> or <\code{\link[rlang:args_data_masking]{data-masking}}> expressions.}
}
\value{
\code{df} with re-positioned columns.
}
\description{
Re-position multiple data frame columns at once.
Extended wrapper around \code{\link[dplyr:relocate]{dplyr::relocate()}} to support relocation of multiple columns to multiple respective positions at once.
}
\details{
All column (selection) indices are converted to names, to allow multiple relocation operations for the same column without index confusion.
}
\examples{
df <- data.frame("a" = 1:2, "b" = 3:4, "c" = 5:6, "d" = 7:8)
df

# Move column "a" after "b" and column "c" after "d"
relocate_multi(df, move = c("a", "c"), after = c("b", "d"))

# The same operation with indices
relocate_multi(df, move = c(1, 3), after = c(2, 4))

# The same operation with tidy-select expression
df |> relocate_multi(move = tidyselect::all_of(c(1, 3)), after = tidyselect::all_of(c(2, 4)))

# Use data masking to move "b" before "a"
df |> relocate_multi(move = rlang::expr(b), before = rlang::expr(a))

# Relocation operations are sequential:
# Move "c" after "d" -> abdc
# Move "a" after "c" -> bdca
# Move "a" after "d" -> bdac
relocate_multi(df, move = c("c", "a", "a"), after = c("d", "c", "d"))

}
