% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_html.R
\name{remove_html}
\alias{remove_html}
\title{Strip html tags}
\usage{
remove_html(vec, squish = TRUE)
}
\arguments{
\item{vec}{Character vector where html tags shall be removed from.}

\item{squish}{If excessive newlines and whitespace shall be removed from stripped text. Uses \code{\link[stringr:str_trim]{stringr::str_squish()}}.}
}
\value{
A character vector without html tags.
}
\description{
Strip html tags from vec and potentially \code{\link[stringr:str_trim]{stringr::str_squish()}} to remove excessive newlines and whitespace.
}
\examples{
text <- "text with <p> some tags </p>"
remove_html(text)
}
