% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_c_na.R
\name{str_c_na}
\alias{str_c_na}
\title{NA-aware str_c}
\usage{
str_c_na(..., sep = "", collapse = NULL)
}
\arguments{
\item{...}{One or more character vectors.

\code{NULL}s are removed; scalar inputs (vectors of length 1) are recycled to
the common length of vector inputs.

Like most other R functions, missing values are "infectious": whenever
a missing value is combined with another string the result will always
be missing. Use \code{\link[dplyr:coalesce]{dplyr::coalesce()}} or \code{\link[stringr:str_replace_na]{str_replace_na()}} to convert to
the desired value.}

\item{sep}{String to insert between input vectors.}

\item{collapse}{Optional string used to combine output into single
string. Generally better to use \code{\link[stringr:str_flatten]{str_flatten()}} if you needed this
behaviour.}
}
\value{
If \code{collapse = NULL} (the default) a character vector with
length equal to the longest input. If \code{collapse} is a string, a character
vector of length 1.
}
\description{
NA-aware drop-in replacement for \link[stringr:str_c]{str_c()}.
Discards NAs when other strings are present. Reduces multiple NAs to a single one.
}
\examples{
vec <- c("a", NA, "b")

# only returns NA
stringr::str_c(vec)

# discards NA
str_c_na(vec)

# collapse also discards NAs
str_c_na(c("a", NA), c(NA, "b"), collapse = ", ")

}
