% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_outlier.R
\name{summarize_outlier}
\alias{summarize_outlier}
\title{Summarize outlier values for numeric vectors}
\usage{
summarize_outlier(
  vec,
  minimal_unique_values = 10,
  method = c("iqr", "zscore_robust"),
  threshold = list(zscore_robust = 7, iqr = 5)
)
}
\arguments{
\item{vec}{The vector to be checked for outliers.}

\item{minimal_unique_values}{Amount of unique values in \code{vec} before the outlier detection is performed. Set to \code{1} to always perform detection.}

\item{method}{The outlier detection method(s). Can be \code{"all"} or some of
\code{"cook"}, \code{"pareto"}, \code{"zscore"}, \code{"zscore_robust"}, \code{"iqr"}, \code{"ci"}, \code{"eti"},
\code{"hdi"}, \code{"bci"}, \code{"mahalanobis"}, \code{"mahalanobis_robust"}, \code{"mcd"}, \code{"ics"},
\code{"optics"} or \code{"lof"}.}

\item{threshold}{A list containing the threshold values for each method (e.g.
\code{list('mahalanobis' = 7, 'cook' = 1)}), above which an observation is
considered as outlier. If \code{NULL}, default values will be used (see
'Details'). If a numeric value is given, it will be used as the threshold for
any of the method run. For EFA/PCA/Omega, indicates the threshold for
correlation of residuals (by default, 0.05).}
}
\value{
A summary string of outliers, an empty string if not outliers where detected.
}
\description{
Utilize \link[performance:check_outliers]{check_outliers()} to determine which values are determined as outliers.
Summarizes the result to a comma-separated, decreasing string.
Allows to specify a minimal amount of unique values in \code{vec} to perform outlier detection.
}
\examples{
norm <- rnorm(n = 100)
shift_norm <- rnorm(n = 10, mean = 999)
vec <- sample(c(norm, shift_norm))

summarize_outlier(vec)
}
