% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_table.R
\name{summarize_table}
\alias{summarize_table}
\title{Summarize metrics of a data.frame or tibble}
\usage{
summarize_table(
  df,
  custom_funs = list(),
  find_culprits = TRUE,
  find_outliers = TRUE,
  find_outlier_opts = list()
)
}
\arguments{
\item{df}{A 'data.frame' of 'tibble'.}

\item{custom_funs}{A named list of custom summary functions.}

\item{find_culprits}{Shall culprit detection be performed via \code{\link[=find_culprit]{find_culprit()}}?}

\item{find_outliers}{Shall outlier detection be performed via \code{\link[=summarize_outlier]{summarize_outlier()}}?}

\item{find_outlier_opts}{If \code{find_outliers}, specify a named list of non-default options for the \code{\link[=summarize_outlier]{summarize_outlier()}} call.}
}
\value{
A summary tibble with as many rows as \code{df} has columns.
}
\description{
Summarize metric for every column in \code{df}.
Collects basic metrics, but can also be extend by custom summary functions (\code{custom_funs}).
Optionally detects \link[=find_culprit]{culprits} or \link[=summarize_outlier]{outliers}.
}
\examples{
# TODO
}
