% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_matching.R
\name{suppress_matching}
\alias{suppress_matching}
\title{Selectively suppress messages, warnings, cat and print}
\usage{
suppress_matching(
  expr,
  pattern = NULL,
  suppress_messages = TRUE,
  suppress_warnings = TRUE,
  suppress_cat_print = TRUE,
  ...
)
}
\arguments{
\item{expr}{An expression that potentially produces some output.}

\item{pattern}{Provide a regex-string to only suppress matching output. If \code{NULL}, suppresses all output.}

\item{suppress_messages}{If output from \code{\link[=message]{message()}} shall be suppressed.}

\item{suppress_warnings}{If output from \code{\link[=warning]{warning()}} shall be suppressed.}

\item{suppress_cat_print}{If output from \code{\link[=cat]{cat()}} and \code{\link[=print]{print()}} shall be suppressed.}

\item{...}{
  Arguments passed on to \code{\link[base:grep]{base::grepl}}
  \describe{
    \item{\code{ignore.case}}{logical. if \code{FALSE}, the pattern matching is \emph{case
      sensitive} and if \code{TRUE}, case is ignored during matching.}
    \item{\code{perl}}{logical.  Should Perl-compatible regexps be used?}
    \item{\code{fixed}}{logical. If \code{TRUE}, \code{pattern} is a string to be
    matched as is.  Overrides all conflicting arguments.}
    \item{\code{useBytes}}{logical.  If \code{TRUE} the matching is done
    byte-by-byte rather than character-by-character.  See
    \sQuote{Details}.}
  }}
}
\value{
The return value of the \code{expr}.
}
\description{
Selectively suppress messages, warnings, cat and/or print whose output matches given regex pattern.
}
\details{
Utilizes \code{\link[=grepl]{grepl()}} for matching output with given regex pattern.
Selective suppression of \code{\link[=cat]{cat()}} and \code{\link[=print]{print()}} is tricky, as differentiation of single output is not trivial.
This function applies a simple heuristic to at least separate \code{\link[=print]{print()}} output from each other and \code{\link[=cat]{cat()}}.
Two consecutive \code{\link[=cat]{cat()}} outputs can not be differentiated at the moment, potentially leading to unwanted selective suppression behavior.
}
\examples{
noisy_fun <- function() {
  message("I am a message")
  warning("I am a warning")
  cat("I am a cat")
  print("I am a print")
  return(TRUE)
}

# suppress all
result <- suppress_matching(noisy_fun())

# suppress by matching regex
result <- suppress_matching(noisy_fun(), pattern = ".*am A.*", ignore.case = TRUE)
}
