% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_t.R
\name{tidy_t}
\alias{tidy_t}
\title{Transpose a data.frame or tibble}
\usage{
tidy_t(
  df,
  id_col = 1,
  exclude_cols = NULL,
  new_id_colname = NULL,
  reguess_coltypes = FALSE,
  name_repair = "unique",
  verbose = FALSE
)
}
\arguments{
\item{df}{The to-be-transposed data.frame or tibble.}

\item{id_col}{Which column of \code{df} shall be used as new column names? Can be either an index or a column name.}

\item{exclude_cols}{Exclude specified columns before transposing. Can be either one or multiple indices or a column names, or \code{NULL}.}

\item{new_id_colname}{Rename the first column of the transposed object. If \code{NULL}, carry over the selected \code{id_col} column name.}

\item{reguess_coltypes}{Shall a re-guessing of column types be performed after transposing?
Allows to correctly retrieve e.g. numeric or logical columns by utilizing \code{\link[=reguess_coltypes]{reguess_coltypes()}}.}

\item{name_repair}{Choose strategy to potentially repair 'id_col', e.g. for uniqueness. Is passed on to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}, supporting among else the following strategies:
\itemize{
\item "minimal": No name repair or checks, beyond basic existence of names and replacing NA's.
\item "unique": Make sure names are unique and not empty. A suffix is appended to duplicate names to make them unique.
\item "universal": Make the names unique and syntactic, meaning that you can safely use the names as variables without causing a syntax error.
}}

\item{verbose}{Set verbosity of potential name_repair.}
}
\value{
A transposed object - keeps object type of \code{df} input (data.frame or tibble).
}
\description{
Allow transposing a data.frame or tibble in a "tidy" way.
Instead of row names, the column names of the transposed object are taken from a specified \code{id_col}.
\cr\strong{Note}: The input \code{df} should \strong{not contain} row names, as they would be discarded.
}
\examples{

df <- data.frame(a = c("x1", "1", "FALSE"), b = c("x2", "2", "TRUE"))

tidy_t(df)

# reguessing column types
tidy_t(df, reguess_coltypes = TRUE)

# using second column as id column
tidy_t(df, id_col = "b")

# excluding columns
tidy_t(df, exclude_cols = "b")

# renaming id column
tidy_t(df, new_id_colname = "var")
}
