% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_dates.R
\name{try_dates}
\alias{try_dates}
\title{Try conversion to date representation}
\usage{
try_dates(vec, ...)
}
\arguments{
\item{vec}{Vector containing values to try conversion.}

\item{...}{
  Arguments passed on to \code{\link[=try_date]{try_date}}
  \describe{
    \item{\code{origin}}{The origin of the serial value, i.e. the day 0. See description for list of supported options.}
    \item{\code{min_date}}{Set a minimum date for conversion. Use format \code{YYYY-MM-DD}.}
    \item{\code{max_date}}{Set a maximum date for conversion. Use format \code{YYYY-MM-DD}.}
  }}
}
\value{
A character vector with potentially converted dates.
}
\description{
Wrapper around \code{\link[=try_date]{try_date()}} for multiple values.
Try conversion of date serial values to their respective date representation.
}
\details{
\strong{Origin}: Correct conversion relies on the correct origin for the serial value.
Available options are:
\itemize{
\item \code{excel}: Set the origin to \code{1899-12-30}. This is the default date origin in Excel for Windows and for Mac since 2011.
\item \code{excel_off1}: Set the origin to \code{1899-12-31}. This might be useful to tackle Excel's 1900 Leap Year Bug for dates before before the 28th of February 1900.
\item \code{excel_1904}: Set the origin to \code{1904-01-01}. This is the date origin for the Excel 1904 date system, which was default in early versions of Excel for Mac.
\item \code{unix}: Set the origin to \code{1970-01-01}. This is Epoch time used by unix-based systems.
}

For more details about the date origin used in Excel, see \href{https://support.microsoft.com/en-us/office/date-systems-in-excel-e7fe7167-48a9-4b96-bb53-5612a800b487}{Microsoft}.
}
\examples{
serial_values <- c(10000, 20000, 30000)

# convert with minimal date set - first value is not converted
try_dates(serial_values, min_date = "1950-01-01")

}
\seealso{
\code{\link[=try_date]{try_date()}}
}
