% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_numeric.R
\name{try_numeric}
\alias{try_numeric}
\title{Try conversion to numeric}
\usage{
try_numeric(x)
}
\arguments{
\item{x}{A single atomic value, i.e. one string, number or element.}
}
\value{
Potentially the numeric representation of \code{x}, \code{x} otherwise.
}
\description{
Try conversion of input element to numeric. Return the numeric representation if it succeeds, the initial value if it does not.
}
\examples{

# convert string representation of a number
try_numeric("5")
try_numeric("1e5")
try_numeric("Inf")

# also converts logicals
try_numeric(TRUE)

# does not alter elements without proper representation
try_numeric("one")
try_numeric(NA)
try_numeric(NULL)

}
