% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_hidra.R
\name{connect_hidra}
\alias{connect_hidra}
\title{Connect to the HIDRA database}
\usage{
connect_hidra(system = "mac")
}
\arguments{
\item{system}{'mac' or 'windows'}
}
\value{
Creates a database connection
}
\description{
Requires \href{https://learn.microsoft.com/en-us/sql/connect/odbc/linux-mac/install-microsoft-odbc-driver-sql-server-macos?view=sql-server-ver16}{ODBC Driver 17 for SQL Server}
if on macOS.
}
\examples{
 \dontrun{con <- connect_hidra("mac")}
}
